requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.5
httpx>=0.27.0
pydantic>=2.6
pydantic-settings>=2.4
PyYAML>=6.0
google-auth>=2.30
gspread>=6.0
python-dotenv>=1.0
orjson>=3.10
